+++ packages/components/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @common-utils/components
 
+## 1.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @common-utils/utils@2.0.2
+
+## 1.1.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @common-utils/utils@2.0.1
+
 ## 1.1.0
 
 ### Minor Changes